facer/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy<2.0.0
scikit-learn>=1.3.0
mediapipe>=0.10.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
validators>=0.22.0
matplotlib>=3.7.0
colour-science>=0.4.2
setuptools>=68.0.0
google-search-results>=2.4.0
scikit-image>=0.21.0
psutil>=5.9.0